Moved change password, update profile and reg form to Forms folder
