+++ components/Forms/RegistrationForm/__tests__/RegistrationForm.test.js
