+++ components/Forms/UpdateProfileForm/UpdateProfileForm.js
@@ -4,7 +4,7 @@ import Router from 'next/router';
 import { getServerErrorMessage } from 'common/utils/api-utils';
 import { array, objectOf, oneOfType, string, number, bool } from 'prop-types';
 import { insertIf } from '@innocuous/functions';
-import MultiStepForm from 'components/Form/MultiStepForm';
+import { MultiStepForm } from '../../Form/MultiStepForm';
 import { ProfessionalDetails, MilitaryStatus, MilitaryDetails, Technology } from './steps';
 
 UpdateProfileForm.propTypes = {
