+++ pages/profile/update.js
@@ -4,7 +4,7 @@ import get from 'lodash/get';
 import Head from 'components/head';
 import HeroBanner from 'components/HeroBanner/HeroBanner';
 import Content from 'components/Content/Content';
-import UpdateProfileForm from 'components/UpdateProfileForm/UpdateProfileForm';
+import UpdateProfileForm from 'components/Forms/UpdateProfileForm/UpdateProfileForm';
 import withAuthSync from 'decorators/withAuthSync/withAuthSync';
 import { getUserPromise } from 'common/constants/api';
